requirements.txt
asgiref==3.2.7
Django==3.0.4
django-bootstrap-breadcrumbs==0.9.2
django-model-utils==4.0.0
Pillow==7.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
